add conditional sentence if the Product is not found.

+++ src/main/java/pws/a/learningRESTful/Controller.java
@@ -47,6 +47,7 @@ public class Controller {
     
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
+        if(!productRepo.containsKey(id))throw new ProductNotfoundException();
         productRepo.remove(id);
         product.setId(id);
         productRepo.put(id, product);
